// src/prisma/generated/models/Wallet.ts
export type WalletUpdateInput = {
  id?: Prisma.StringFieldUpdateOperationsInput | string
  walletType?: Prisma.EnumWalletTypeFieldUpdateOperationsInput | $Enums.WalletType
  balance?: Prisma.FloatFieldUpdateOperationsInput | number
  academy?: Prisma.AcademyUpdateOneRequiredWithoutWalletNestedInput
  client?: Prisma.ClientUpdateOneWithoutWalletNestedInput
  jobProfile?: Prisma.JobProfileUpdateOneWithoutWalletsNestedInput
  receivers?: Prisma.WalletMovementUpdateManyWithoutReceiverNestedInput
  senders?: Prisma.WalletMovementUpdateManyWithoutSenderNestedInput
}
// ...
export type EnumWalletTypeFieldUpdateOperationsInput = {
  set?: $Enums.WalletType
}

export type FloatFieldUpdateOperationsInput = {
  set?: number
  increment?: number
  decrement?: number
  multiply?: number
  divide?: number
}